yum: Ensure we disable the MediaRepo when the PackageKit backend has finished
YUM is unable to ignore repos that do not exist, even media repos
and this stops yum working if the media is not present.

PackageKit is a bit more clever and can ignore (with a warning)
sources that do not exist or are not contactable but we still need
to preserve compatibility for people who use both the PackageKit
tools and the yum CLI.

Ensure we set the enabled= line in the repo file so that any tool
that cannot skip repos is able to continue working.